requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.48
botocore==1.40.48
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
curl_cffi==0.13.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
fastapi==0.118.2
Flask==3.1.2
flask-cors==6.0.1
frozendict==2.4.6
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langsmith==0.4.33
MarkupSafe==3.0.3
mcp==1.16.0
multidict==6.7.0
multitasking==0.0.12
nanda-adapter==1.0.1
numpy==2.3.3
openai==2.2.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.32.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pymongo==4.15.3
python-a2a==0.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
yfinance==0.2.66
zstandard==0.25.0
